change float to double

+++ src/main/java/com/asia/compiler/common/utils/Type.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum Type {
     INT("i32"),
-    FLOAT("float"),
+    FLOAT("double"),
     STRING("%String");
 
     @Getter
